refactor(registry): shorten tools.jamf_pro namespace to tools.jamf
Rename the namespace and every identifier coupled to it so the family reads
consistently: the secret becomes jamf/JAMF_API_TOKEN, the VARS key becomes
VARS.jamf, and the OAuth provider id becomes jamf, which shortens its derived
token to JAMF_SERVICE_TOKEN.

The display group stays 'Jamf Pro' because that is the product name, following
the same split already used for tools.iru with 'Iru (Kandji)'.

Nothing here is released yet, so no stored integration or workspace secret
refers to the old identifiers. The JAMF_PRO_* stdio variables in the private
MCP catalog belong to the separate Jamf MCP hub and are untouched.
