+++ frontend/src/components/icons.tsx
@@ -659,7 +659,7 @@ export const UDFIcons: Record<string, (props: CustomIconProps) => JSX.Element> =
     "tools.elastic_security": createCatalogIconRenderer("elastic_mcp"),
     "tools.elasticsearch": createCatalogIconRenderer("elastic_mcp"),
     "tools.iru": createCatalogIconRenderer("iru_mcp"),
-    "tools.jamf_pro": createCatalogIconRenderer("jamf_mcp"),
+    "tools.jamf": createCatalogIconRenderer("jamf_mcp"),
     "tools.panther": createCatalogIconRenderer("panther_mcp"),
     "tools.rootly": createCatalogIconRenderer("rootly_mcp"),
     "tools.sentinel_one": createCatalogIconRenderer("sentinelone_mcp"),
@@ -865,7 +865,7 @@ export const providerIcons: Record<
       <ServiceNowIcon {...rest} />
     </div>
   ),
-  jamf_pro: createCatalogIconRenderer("jamf_mcp"),
+  jamf: createCatalogIconRenderer("jamf_mcp"),
   slack: ({ className, iconClassName, flairsize: _ignored, ...rest }) => (
     <div className={cn("!rounded-sm", className)}>
       <SlackIcon {...rest} className={cn("size-full", iconClassName)} />
@@ -978,7 +978,7 @@ export const secretIcons: Record<
       <JiraIcon {...rest} />
     </div>
   ),
-  jamf_pro: createCatalogIconRenderer("jamf_mcp"),
+  jamf: createCatalogIconRenderer("jamf_mcp"),
   iru: createCatalogIconRenderer("iru_mcp"),
   datadog: ({ className, ...rest }) => (
     <div className={className}>
