+++ packages/tracecat-registry/tracecat_registry/templates/tools/jamf/computers/erase_computer.yml
@@ -1,0 +1,49 @@
+type: action
+definition:
+  title: Erase computer
+  description: Irreversibly erase a Jamf Pro computer, optionally supplying command options such as a PIN.
+  display_group: Jamf Pro
+  doc_url: https://developer.jamf.com/jamf-pro/reference/post_v1-computer-inventory-id-erase
+  namespace: tools.jamf
+  name: erase_computer
+  secrets:
+    - type: oauth
+      provider_id: jamf
+      grant_type: client_credentials
+      optional: true
+    - name: jamf
+      keys: ["JAMF_API_TOKEN"]
+      optional: true
+  expects:
+    base_url:
+      type: str | None
+      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
+      default: null
+    computer_id:
+      type: str
+      description: Computer ID.
+    payload:
+      type: dict[str, any] | None
+      description: Optional erase command fields such as pin.
+      default: null
+  steps:
+    - ref: build_payload
+      action: core.script.run_python
+      args:
+        inputs:
+          payload: ${{ inputs.payload }}
+        script: |
+          def main(payload):
+              if payload is None:
+                  return {}
+              return {k: v for k, v in payload.items() if v is not None}
+    - ref: erase_computer
+      action: core.http_request
+      args:
+        url: ${{ inputs.base_url || VARS.jamf.base_url }}/api/v1/computer-inventory/${{ FN.url_encode(inputs.computer_id) }}/erase
+        method: POST
+        headers:
+          Authorization: Bearer ${{ SECRETS.jamf_oauth.JAMF_SERVICE_TOKEN || SECRETS.jamf.JAMF_API_TOKEN }}
+          Content-Type: application/json
+        payload: ${{ steps.build_payload.result }}
+  returns: ${{ steps.erase_computer.result }}
