+++ packages/tracecat-registry/tracecat_registry/templates/tools/jamf/computers/get_computer_inventory_detail.yml
@@ -1,0 +1,33 @@
+type: action
+definition:
+  title: Get computer inventory detail
+  description: Get detailed inventory for a Jamf Pro computer.
+  display_group: Jamf Pro
+  doc_url: https://developer.jamf.com/jamf-pro/reference/get_v1-computers-inventory-detail-id
+  namespace: tools.jamf
+  name: get_computer_inventory_detail
+  secrets:
+    - type: oauth
+      provider_id: jamf
+      grant_type: client_credentials
+      optional: true
+    - name: jamf
+      keys: ["JAMF_API_TOKEN"]
+      optional: true
+  expects:
+    base_url:
+      type: str | None
+      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
+      default: null
+    computer_id:
+      type: str
+      description: Computer ID.
+  steps:
+    - ref: get_computer_inventory_detail
+      action: core.http_request
+      args:
+        url: ${{ inputs.base_url || VARS.jamf.base_url }}/api/v1/computers-inventory-detail/${{ FN.url_encode(inputs.computer_id) }}
+        method: GET
+        headers:
+          Authorization: Bearer ${{ SECRETS.jamf_oauth.JAMF_SERVICE_TOKEN || SECRETS.jamf.JAMF_API_TOKEN }}
+  returns: ${{ steps.get_computer_inventory_detail.result }}
