+++ packages/tracecat-registry/tracecat_registry/templates/tools/jamf/computers/redeploy_management_framework.yml
@@ -1,0 +1,33 @@
+type: action
+definition:
+  title: Redeploy Jamf management framework
+  description: Reinstall the Jamf management framework (QuickAdd) on a Mac when the agent is broken; returns 202 with deviceId and commandUuid.
+  display_group: Jamf Pro
+  doc_url: https://developer.jamf.com/jamf-pro/reference/post_v1-jamf-management-framework-redeploy-id
+  namespace: tools.jamf
+  name: redeploy_management_framework
+  secrets:
+    - type: oauth
+      provider_id: jamf
+      grant_type: client_credentials
+      optional: true
+    - name: jamf
+      keys: ["JAMF_API_TOKEN"]
+      optional: true
+  expects:
+    base_url:
+      type: str | None
+      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
+      default: null
+    computer_id:
+      type: str
+      description: Computer ID.
+  steps:
+    - ref: redeploy_management_framework
+      action: core.http_request
+      args:
+        url: ${{ inputs.base_url || VARS.jamf.base_url }}/api/v1/jamf-management-framework/redeploy/${{ FN.url_encode(inputs.computer_id) }}
+        method: POST
+        headers:
+          Authorization: Bearer ${{ SECRETS.jamf_oauth.JAMF_SERVICE_TOKEN || SECRETS.jamf.JAMF_API_TOKEN }}
+  returns: ${{ steps.redeploy_management_framework.result }}
