+++ packages/tracecat-registry/tracecat_registry/templates/tools/jamf/computers/remove_computer_mdm_profile.yml
@@ -1,0 +1,33 @@
+type: action
+definition:
+  title: Remove computer MDM profile
+  description: Remove the MDM profile from a Jamf Pro computer.
+  display_group: Jamf Pro
+  doc_url: https://developer.jamf.com/jamf-pro/reference/post_v1-computer-inventory-id-remove-mdm-profile
+  namespace: tools.jamf
+  name: remove_computer_mdm_profile
+  secrets:
+    - type: oauth
+      provider_id: jamf
+      grant_type: client_credentials
+      optional: true
+    - name: jamf
+      keys: ["JAMF_API_TOKEN"]
+      optional: true
+  expects:
+    base_url:
+      type: str | None
+      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
+      default: null
+    computer_id:
+      type: str
+      description: Computer ID.
+  steps:
+    - ref: remove_computer_mdm_profile
+      action: core.http_request
+      args:
+        url: ${{ inputs.base_url || VARS.jamf.base_url }}/api/v1/computer-inventory/${{ FN.url_encode(inputs.computer_id) }}/remove-mdm-profile
+        method: POST
+        headers:
+          Authorization: Bearer ${{ SECRETS.jamf_oauth.JAMF_SERVICE_TOKEN || SECRETS.jamf.JAMF_API_TOKEN }}
+  returns: ${{ steps.remove_computer_mdm_profile.result }}
