+++ packages/tracecat-registry/tracecat_registry/templates/tools/jamf/groups/get_smart_group_membership.yml
@@ -1,0 +1,33 @@
+type: action
+definition:
+  title: Get smart computer group membership
+  description: Get the computer membership of a Jamf Pro smart group.
+  display_group: Jamf Pro
+  doc_url: https://developer.jamf.com/jamf-pro/reference/get_v3-computer-groups-smart-group-membership-id
+  namespace: tools.jamf
+  name: get_smart_group_membership
+  secrets:
+    - type: oauth
+      provider_id: jamf
+      grant_type: client_credentials
+      optional: true
+    - name: jamf
+      keys: ["JAMF_API_TOKEN"]
+      optional: true
+  expects:
+    base_url:
+      type: str | None
+      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
+      default: null
+    group_id:
+      type: str
+      description: Smart computer group ID.
+  steps:
+    - ref: get_smart_group_membership
+      action: core.http_request
+      args:
+        url: ${{ inputs.base_url || VARS.jamf.base_url }}/api/v3/computer-groups/smart-group-membership/${{ FN.url_encode(inputs.group_id) }}
+        method: GET
+        headers:
+          Authorization: Bearer ${{ SECRETS.jamf_oauth.JAMF_SERVICE_TOKEN || SECRETS.jamf.JAMF_API_TOKEN }}
+  returns: ${{ steps.get_smart_group_membership.result }}
