+++ packages/tracecat-registry/tracecat_registry/templates/tools/jamf/groups/list_computer_groups.yml
@@ -1,0 +1,30 @@
+type: action
+definition:
+  title: List computer groups
+  description: List Jamf Pro computer groups.
+  display_group: Jamf Pro
+  doc_url: https://developer.jamf.com/jamf-pro/reference/get_v1-computer-groups
+  namespace: tools.jamf
+  name: list_computer_groups
+  secrets:
+    - type: oauth
+      provider_id: jamf
+      grant_type: client_credentials
+      optional: true
+    - name: jamf
+      keys: ["JAMF_API_TOKEN"]
+      optional: true
+  expects:
+    base_url:
+      type: str | None
+      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
+      default: null
+  steps:
+    - ref: list_computer_groups
+      action: core.http_request
+      args:
+        url: ${{ inputs.base_url || VARS.jamf.base_url }}/api/v1/computer-groups
+        method: GET
+        headers:
+          Authorization: Bearer ${{ SECRETS.jamf_oauth.JAMF_SERVICE_TOKEN || SECRETS.jamf.JAMF_API_TOKEN }}
+  returns: ${{ steps.list_computer_groups.result }}
