--- packages/tracecat-registry/tracecat_registry/templates/tools/jamf_pro/commands/send_mdm_command.yml
@@ -1,52 +1,0 @@
-type: action
-definition:
-  title: Send MDM command
-  description: Send an MDM command to one or more Jamf Pro clients; returns 201 with each command ID and href.
-  display_group: Jamf Pro
-  doc_url: https://developer.jamf.com/jamf-pro/reference/post_v2-mdm-commands
-  namespace: tools.jamf_pro
-  name: send_mdm_command
-  secrets:
-    - type: oauth
-      provider_id: jamf_pro
-      grant_type: client_credentials
-      optional: true
-    - name: jamf_pro
-      keys: ["JAMF_PRO_API_TOKEN"]
-      optional: true
-  expects:
-    base_url:
-      type: str | None
-      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
-      default: null
-    client_data:
-      type: list[dict[str, any]]
-      description: Target clients, with each entry shaped as {"managementId":"<uuid>"}.
-    command_data:
-      type: dict[str, any]
-      description: >-
-        Command fields including commandType. Accepted command types:
-        APPLY_REDEMPTION_CODE, CANCEL_ENHANCED_LOG_COLLECTION, CERTIFICATE_LIST,
-        CLEAR_PASSCODE, CLEAR_RESTRICTIONS_PASSWORD, DECLARATIVE_MANAGEMENT,
-        DELETE_USER, DEVICE_INFORMATION, DEVICE_LOCATION, DEVICE_LOCK,
-        DISABLE_LOST_MODE, DISABLE_REMOTE_DESKTOP, ENABLE_LOST_MODE,
-        ENABLE_REMOTE_DESKTOP, ERASE_DEVICE, INSTALLED_APPLICATION_LIST,
-        LOG_OUT_USER, MANAGED_APPLICATION_LIST, MANAGED_MEDIA_LIST,
-        REFRESH_CELLULAR_PLANS, PLAY_LOST_MODE_SOUND, PROFILE_LIST,
-        PROVISIONING_PROFILE_LIST, RESTART_DEVICE, REQUEST_MIRRORING,
-        SECURITY_INFO, SETTINGS, SET_AUTO_ADMIN_PASSWORD, SET_RECOVERY_LOCK,
-        SHUT_DOWN_DEVICE, STOP_MIRRORING, TRIGGER_ENHANCED_LOG_COLLECTION,
-        UNLOCK_USER_ACCOUNT, VALIDATE_APPLICATIONS, VERIFY_RECOVERY_LOCK.
-  steps:
-    - ref: send_mdm_command
-      action: core.http_request
-      args:
-        url: ${{ inputs.base_url || VARS.jamf_pro.base_url }}/api/v2/mdm/commands
-        method: POST
-        headers:
-          Authorization: Bearer ${{ SECRETS.jamf_pro_oauth.JAMF_PRO_SERVICE_TOKEN || SECRETS.jamf_pro.JAMF_PRO_API_TOKEN }}
-          Content-Type: application/json
-        payload:
-          clientData: ${{ inputs.client_data }}
-          commandData: ${{ inputs.command_data }}
-  returns: ${{ steps.send_mdm_command.result }}
