--- packages/tracecat-registry/tracecat_registry/templates/tools/jamf_pro/computers/delete_computer.yml
@@ -1,33 +1,0 @@
-type: action
-definition:
-  title: Delete computer
-  description: Delete only the computer inventory record from Jamf Pro without erasing the computer.
-  display_group: Jamf Pro
-  doc_url: https://developer.jamf.com/jamf-pro/reference/delete_v1-computers-inventory-id
-  namespace: tools.jamf_pro
-  name: delete_computer
-  secrets:
-    - type: oauth
-      provider_id: jamf_pro
-      grant_type: client_credentials
-      optional: true
-    - name: jamf_pro
-      keys: ["JAMF_PRO_API_TOKEN"]
-      optional: true
-  expects:
-    base_url:
-      type: str | None
-      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
-      default: null
-    computer_id:
-      type: str
-      description: Computer ID.
-  steps:
-    - ref: delete_computer
-      action: core.http_request
-      args:
-        url: ${{ inputs.base_url || VARS.jamf_pro.base_url }}/api/v1/computers-inventory/${{ FN.url_encode(inputs.computer_id) }}
-        method: DELETE
-        headers:
-          Authorization: Bearer ${{ SECRETS.jamf_pro_oauth.JAMF_PRO_SERVICE_TOKEN || SECRETS.jamf_pro.JAMF_PRO_API_TOKEN }}
-  returns: ${{ steps.delete_computer.result }}
