--- packages/tracecat-registry/tracecat_registry/templates/tools/jamf_pro/computers/get_computer_recovery_lock_password.yml
@@ -1,33 +1,0 @@
-type: action
-definition:
-  title: Get computer recovery lock password
-  description: Get the recovery lock password for a Jamf Pro computer.
-  display_group: Jamf Pro
-  doc_url: https://developer.jamf.com/jamf-pro/reference/get_v1-computers-inventory-id-view-recovery-lock-password
-  namespace: tools.jamf_pro
-  name: get_computer_recovery_lock_password
-  secrets:
-    - type: oauth
-      provider_id: jamf_pro
-      grant_type: client_credentials
-      optional: true
-    - name: jamf_pro
-      keys: ["JAMF_PRO_API_TOKEN"]
-      optional: true
-  expects:
-    base_url:
-      type: str | None
-      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
-      default: null
-    computer_id:
-      type: str
-      description: Computer ID.
-  steps:
-    - ref: get_computer_recovery_lock_password
-      action: core.http_request
-      args:
-        url: ${{ inputs.base_url || VARS.jamf_pro.base_url }}/api/v1/computers-inventory/${{ FN.url_encode(inputs.computer_id) }}/view-recovery-lock-password
-        method: GET
-        headers:
-          Authorization: Bearer ${{ SECRETS.jamf_pro_oauth.JAMF_PRO_SERVICE_TOKEN || SECRETS.jamf_pro.JAMF_PRO_API_TOKEN }}
-  returns: ${{ steps.get_computer_recovery_lock_password.result }}
