--- packages/tracecat-registry/tracecat_registry/templates/tools/jamf_pro/computers/redeploy_management_framework.yml
@@ -1,33 +1,0 @@
-type: action
-definition:
-  title: Redeploy Jamf management framework
-  description: Reinstall the Jamf management framework (QuickAdd) on a Mac when the agent is broken; returns 202 with deviceId and commandUuid.
-  display_group: Jamf Pro
-  doc_url: https://developer.jamf.com/jamf-pro/reference/post_v1-jamf-management-framework-redeploy-id
-  namespace: tools.jamf_pro
-  name: redeploy_management_framework
-  secrets:
-    - type: oauth
-      provider_id: jamf_pro
-      grant_type: client_credentials
-      optional: true
-    - name: jamf_pro
-      keys: ["JAMF_PRO_API_TOKEN"]
-      optional: true
-  expects:
-    base_url:
-      type: str | None
-      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
-      default: null
-    computer_id:
-      type: str
-      description: Computer ID.
-  steps:
-    - ref: redeploy_management_framework
-      action: core.http_request
-      args:
-        url: ${{ inputs.base_url || VARS.jamf_pro.base_url }}/api/v1/jamf-management-framework/redeploy/${{ FN.url_encode(inputs.computer_id) }}
-        method: POST
-        headers:
-          Authorization: Bearer ${{ SECRETS.jamf_pro_oauth.JAMF_PRO_SERVICE_TOKEN || SECRETS.jamf_pro.JAMF_PRO_API_TOKEN }}
-  returns: ${{ steps.redeploy_management_framework.result }}
