--- packages/tracecat-registry/tracecat_registry/templates/tools/jamf_pro/computers/update_computer_inventory_detail.yml
@@ -1,38 +1,0 @@
-type: action
-definition:
-  title: Update computer inventory detail
-  description: Partially update computer inventory detail; use the userAndLocation section to assign the device to a user.
-  display_group: Jamf Pro
-  doc_url: https://developer.jamf.com/jamf-pro/reference/patch_v1-computers-inventory-detail-id
-  namespace: tools.jamf_pro
-  name: update_computer_inventory_detail
-  secrets:
-    - type: oauth
-      provider_id: jamf_pro
-      grant_type: client_credentials
-      optional: true
-    - name: jamf_pro
-      keys: ["JAMF_PRO_API_TOKEN"]
-      optional: true
-  expects:
-    base_url:
-      type: str | None
-      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
-      default: null
-    computer_id:
-      type: str
-      description: Computer ID.
-    payload:
-      type: dict[str, any]
-      description: Partial inventory update; the userAndLocation section assigns the device to a user.
-  steps:
-    - ref: update_computer_inventory_detail
-      action: core.http_request
-      args:
-        url: ${{ inputs.base_url || VARS.jamf_pro.base_url }}/api/v1/computers-inventory-detail/${{ FN.url_encode(inputs.computer_id) }}
-        method: PATCH
-        headers:
-          Authorization: Bearer ${{ SECRETS.jamf_pro_oauth.JAMF_PRO_SERVICE_TOKEN || SECRETS.jamf_pro.JAMF_PRO_API_TOKEN }}
-          Content-Type: application/json
-        payload: ${{ inputs.payload }}
-  returns: ${{ steps.update_computer_inventory_detail.result }}
