--- packages/tracecat-registry/tracecat_registry/templates/tools/jamf_pro/enrollment/list_device_enrollment_devices.yml
@@ -1,33 +1,0 @@
-type: action
-definition:
-  title: List automated device enrollment devices
-  description: List devices associated with a Jamf Pro Automated Device Enrollment (ADE/ABM/ASM) instance.
-  display_group: Jamf Pro
-  doc_url: https://developer.jamf.com/jamf-pro/reference/get_v1-device-enrollments-id-devices
-  namespace: tools.jamf_pro
-  name: list_device_enrollment_devices
-  secrets:
-    - type: oauth
-      provider_id: jamf_pro
-      grant_type: client_credentials
-      optional: true
-    - name: jamf_pro
-      keys: ["JAMF_PRO_API_TOKEN"]
-      optional: true
-  expects:
-    base_url:
-      type: str | None
-      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
-      default: null
-    device_enrollment_id:
-      type: str
-      description: Automated Device Enrollment instance ID.
-  steps:
-    - ref: list_device_enrollment_devices
-      action: core.http_request
-      args:
-        url: ${{ inputs.base_url || VARS.jamf_pro.base_url }}/api/v1/device-enrollments/${{ FN.url_encode(inputs.device_enrollment_id) }}/devices
-        method: GET
-        headers:
-          Authorization: Bearer ${{ SECRETS.jamf_pro_oauth.JAMF_PRO_SERVICE_TOKEN || SECRETS.jamf_pro.JAMF_PRO_API_TOKEN }}
-  returns: ${{ steps.list_device_enrollment_devices.result }}
