--- packages/tracecat-registry/tracecat_registry/templates/tools/jamf_pro/mobile_devices/list_mobile_devices.yml
@@ -1,59 +1,0 @@
-type: action
-definition:
-  title: List mobile devices
-  description: List Jamf Pro mobile devices in the standard paged response envelope.
-  display_group: Jamf Pro
-  doc_url: https://developer.jamf.com/jamf-pro/reference/get_v2-mobile-devices
-  namespace: tools.jamf_pro
-  name: list_mobile_devices
-  secrets:
-    - type: oauth
-      provider_id: jamf_pro
-      grant_type: client_credentials
-      optional: true
-    - name: jamf_pro
-      keys: ["JAMF_PRO_API_TOKEN"]
-      optional: true
-  expects:
-    base_url:
-      type: str | None
-      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
-      default: null
-    page:
-      type: int | None
-      description: Zero-indexed page number.
-      default: null
-    page_size:
-      type: int | None
-      description: Number of results per page.
-      default: null
-    sort:
-      type: list[str] | None
-      description: Sort expressions in property:asc or property:desc format. Defaults to id:asc.
-      default: null
-    filter:
-      type: str | None
-      description: RSQL filter expression. See https://developer.jamf.com/jamf-pro/docs/filtering-with-rsql.
-      default: null
-  steps:
-    - ref: build_params
-      action: core.script.run_python
-      args:
-        inputs:
-          params:
-            page: ${{ inputs.page }}
-            page-size: ${{ inputs.page_size }}
-            sort: ${{ inputs.sort }}
-            filter: ${{ inputs.filter }}
-        script: |
-          def main(params):
-              return {k: v for k, v in params.items() if v is not None}
-    - ref: list_mobile_devices
-      action: core.http_request
-      args:
-        url: ${{ inputs.base_url || VARS.jamf_pro.base_url }}/api/v2/mobile-devices
-        method: GET
-        headers:
-          Authorization: Bearer ${{ SECRETS.jamf_pro_oauth.JAMF_PRO_SERVICE_TOKEN || SECRETS.jamf_pro.JAMF_PRO_API_TOKEN }}
-        params: ${{ steps.build_params.result }}
-  returns: ${{ steps.list_mobile_devices.result }}
