--- packages/tracecat-registry/tracecat_registry/templates/tools/jamf_pro/prestage/add_computer_prestage_scope.yml
@@ -1,43 +1,0 @@
-type: action
-definition:
-  title: Add serial numbers to computer PreStage scope
-  description: Add serial numbers to a computer PreStage scope; a 409 is an optimistic-locking conflict, so refetch the scope for the current versionLock and retry.
-  display_group: Jamf Pro
-  doc_url: https://developer.jamf.com/jamf-pro/reference/post_v2-computer-prestages-id-scope
-  namespace: tools.jamf_pro
-  name: add_computer_prestage_scope
-  secrets:
-    - type: oauth
-      provider_id: jamf_pro
-      grant_type: client_credentials
-      optional: true
-    - name: jamf_pro
-      keys: ["JAMF_PRO_API_TOKEN"]
-      optional: true
-  expects:
-    base_url:
-      type: str | None
-      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
-      default: null
-    prestage_id:
-      type: str
-      description: Computer PreStage enrollment ID.
-    serial_numbers:
-      type: list[str]
-      description: Serial numbers to add to the PreStage scope.
-    version_lock:
-      type: int
-      description: Current optimistic-lock version from the PreStage scope.
-  steps:
-    - ref: add_computer_prestage_scope
-      action: core.http_request
-      args:
-        url: ${{ inputs.base_url || VARS.jamf_pro.base_url }}/api/v2/computer-prestages/${{ FN.url_encode(inputs.prestage_id) }}/scope
-        method: POST
-        headers:
-          Authorization: Bearer ${{ SECRETS.jamf_pro_oauth.JAMF_PRO_SERVICE_TOKEN || SECRETS.jamf_pro.JAMF_PRO_API_TOKEN }}
-          Content-Type: application/json
-        payload:
-          serialNumbers: ${{ inputs.serial_numbers }}
-          versionLock: ${{ inputs.version_lock }}
-  returns: ${{ steps.add_computer_prestage_scope.result }}
