--- packages/tracecat-registry/tracecat_registry/templates/tools/jamf_pro/prestage/get_computer_prestage_scope.yml
@@ -1,33 +1,0 @@
-type: action
-definition:
-  title: Get computer PreStage scope
-  description: Get the serial numbers and version lock in a Jamf Pro computer PreStage scope.
-  display_group: Jamf Pro
-  doc_url: https://developer.jamf.com/jamf-pro/reference/get_v2-computer-prestages-id-scope
-  namespace: tools.jamf_pro
-  name: get_computer_prestage_scope
-  secrets:
-    - type: oauth
-      provider_id: jamf_pro
-      grant_type: client_credentials
-      optional: true
-    - name: jamf_pro
-      keys: ["JAMF_PRO_API_TOKEN"]
-      optional: true
-  expects:
-    base_url:
-      type: str | None
-      description: Jamf Pro base URL (e.g. https://acme.jamfcloud.com).
-      default: null
-    prestage_id:
-      type: str
-      description: Computer PreStage enrollment ID.
-  steps:
-    - ref: get_computer_prestage_scope
-      action: core.http_request
-      args:
-        url: ${{ inputs.base_url || VARS.jamf_pro.base_url }}/api/v2/computer-prestages/${{ FN.url_encode(inputs.prestage_id) }}/scope
-        method: GET
-        headers:
-          Authorization: Bearer ${{ SECRETS.jamf_pro_oauth.JAMF_PRO_SERVICE_TOKEN || SECRETS.jamf_pro.JAMF_PRO_API_TOKEN }}
-  returns: ${{ steps.get_computer_prestage_scope.result }}
